Shrimp damages player and can die
