+++ Assets/Prefabs/Player Mermaid.prefab
@@ -9,7 +9,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 7882433414056290704}
-  - component: {fileID: 6936004740129785646}
   - component: {fileID: 2623236400860764829}
   - component: {fileID: 594513264201488557}
   - component: {fileID: 5318723798107836790}
@@ -41,19 +40,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &6936004740129785646
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 329315263273244811}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: c81ce96205127554dae8696ae6710886, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  AnimationController: {fileID: 3391367493163675675}
 --- !u!50 &2623236400860764829
 Rigidbody2D:
   serializedVersion: 4
