+++ Assets/Prefabs/Shrimp Enemy.prefab
@@ -48,6 +48,7 @@ GameObject:
   - component: {fileID: 7430090946755638391}
   - component: {fileID: 4718369574450353398}
   - component: {fileID: 1756635384423341092}
+  - component: {fileID: 1838894602939110643}
   m_Layer: 7
   m_Name: Shrimp
   m_TagString: Untagged
@@ -178,6 +179,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   MaxHealth: 2
   Health: 2
+  PrefabRoot: {fileID: 1682477695393711541}
 --- !u!95 &4718369574450353398
 Animator:
   serializedVersion: 5
@@ -211,6 +213,20 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e745e46e1d4711046bbb3d39ebe8da50, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &1838894602939110643
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2474266779345672690}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 6ac77f2b9fbba184e86417427a46b233, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  DamageValue: 1
+  Cooldown: 1
 --- !u!1 &5617366960630185564
 GameObject:
   m_ObjectHideFlags: 0
