+++ Assets/Scripts/Enemies/Shrimp/ShrimpDeath.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a1545772da3349e4f9768e46a079aece
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
