+++ Assets/Scripts/Player/Projectile.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 548579a5c5e6b0549bb946419cea6040
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
