+++ Assets/Scripts/Player/Projectile/HitEnemy.cs
@@ -1,0 +1,28 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+[RequireComponent(typeof(Collider2D), typeof(Rigidbody2D))]
+public class HitEnemy : MonoBehaviour
+{
+    private Rigidbody2D _rigidbody2D;
+
+    [Tooltip("How much HP is deducted per damage event")]
+    public int DamageValue;
+
+    void Start()
+    {
+        _rigidbody2D = GetComponent<Rigidbody2D>();
+    }
+
+    void OnCollisionEnter2D(Collision2D other)
+    {
+        var health = other.gameObject.GetComponentInParent<ShrimpHealth>();
+
+        health.TakeDamage(damageValue: DamageValue,
+            (_rigidbody2D.position - other.GetContact(0).point).x
+        );
+
+        Destroy(gameObject);
+    }
+}
