+++ Assets/Scripts/Player/Projectile/HitEnemy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be17c4cd880f5eb4d9b3bcb0f1ff451d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
