--- Assets/Scripts/Player/StarfishController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9d378441b2077fc47ba5a11b93ac6b94
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
